requirements_fixed.txt
# Fixed requirements for PerceptionGPT
# Compatible versions tested and working

# Core ML frameworks
torch>=2.0.0,<2.3.0
torchvision>=0.15.0,<0.18.0
torchaudio>=2.0.0,<2.3.0

# Transformers and HuggingFace - FIXED VERSIONS
transformers==4.46.3  # Compatible with torch 2.2.2 and peft 0.13.2
datasets>=2.14.0
accelerate>=0.24.0
peft==0.13.2  # Compatible with transformers 4.46.3

# DeepSpeed for distributed training
deepspeed>=0.12.0

# Computer Vision
pillow>=9.5.0
opencv-python>=4.8.0
pycocotools>=2.0.0

# Data processing
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
nltk>=3.8.0

# Training utilities
mmengine>=0.9.0
tensorboard>=2.14.0
einops>=0.7.0
tqdm>=4.66.0
matplotlib>=3.7.0

# Tokenization
sentencepiece>=0.1.99

# Memory optimization (for 4GB GPU)
bitsandbytes>=0.41.0  # For 8-bit training

# Other utilities
scikit-image>=0.19.0
timm>=0.9.0
protobuf>=3.20.0
